refactor(return weather function)
The function now is not async, it will take an event as an argument to be used to prevent the default behavior of the form submit. The input field and search button are selected. The search button will have an event listener waiting for the click event to then call an async function using all the previous code but this time creating a key word with the input field value to use it in the fetching process.